refactor teleclass again
