+++ autoreg_metadata/common/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Any, Generic, TypeVar
 
-from pydantic import BaseModel, ConfigDict, Field
+from pydantic import BaseModel, Field
 
 # Define a generic type for source-specific data
 T = TypeVar("T")
