+++ autoreg_metadata/grouper/teleclass/classifier/similarity.py
@@ -1,9 +1,7 @@
 import numpy as np
 
 from autoreg_metadata.grouper.teleclass.core.embeddings import EmbeddingService
-from autoreg_metadata.grouper.teleclass.core.models.enrichment_models import (
-    EnrichedClass,
-)
+from autoreg_metadata.grouper.teleclass.core.models import EnrichedClass
 from autoreg_metadata.grouper.teleclass.core.taxonomy_manager import TaxonomyManager
 from autoreg_metadata.log import logger
